Extract CoordinatorAddStudentModal to separate file, modify button colors, reformat
