+++ csm_web/frontend/src/components/MentorSection.js
@@ -1,15 +1,14 @@
 import React, { useState, useEffect } from "react";
-import { Link } from "react-router-dom";
 import PropTypes from "prop-types";
 import { fetchJSON, fetchWithMethod, HTTP_METHODS } from "../utils/api";
 import { SPACETIME_SHAPE } from "../utils/types";
 import { SectionDetail, InfoCard, SectionSpacetime, ROLES } from "./Section";
+import { CoordinatorAddStudentModal } from "./CoordinatorAddStudentModal";
 import { Switch, Route, NavLink } from "react-router-dom";
 import { groupBy } from "lodash";
 import CopyIcon from "../../static/frontend/img/copy.svg";
 import CheckCircle from "../../static/frontend/img/check_circle.svg";
 import PencilIcon from "../../static/frontend/img/pencil.svg";
-import ErrorCircle from "../../static/frontend/img/error_outline.svg";
 import { ATTENDANCE_LABELS } from "./Section";
 import Modal from "./Modal";
 import LoadingSpinner from "./LoadingSpinner";
@@ -723,453 +722,6 @@ MentorSectionInfo.propTypes = {
   capacity: PropTypes.number,
   description: PropTypes.string,
   id: PropTypes.number.isRequired
-};
-
-function CoordinatorAddStudentModal({ closeModal, userEmails, sectionId }) {
-  const [emailsToAdd, setEmailsToAdd] = useState([""]);
-  const [response, setResponse] = useState({});
-  /**
-   * Mapping from email to action for what the user replies with after the first submission
-   */
-  const [responseActions, setResponseActions] = useState(new Map());
-  const [addStage, setAddStage] = useState(CoordinatorAddStudentModal.STATES.INITIAL);
-  /**
-   * Whether or not the form fields should be highlighted if invalid
-   */
-  const [validationEnabled, setValidationEnabled] = useState(false);
-
-  function setEmail(index, value) {
-    let newEmailsToAdd = [...emailsToAdd];
-    newEmailsToAdd[index] = value;
-    setEmailsToAdd(newEmailsToAdd);
-  }
-
-  function addNewEmail() {
-    setEmailsToAdd([...emailsToAdd, ""]);
-  }
-
-  function addEmailsFromClipboard() {
-    // this will error with a DOMException if clipboard access is denied
-    navigator.clipboard.readText().then(text => {
-      setEmailsToAdd(lastEmails => {
-        if (lastEmails.length == 1 && lastEmails[0] === "") {
-          // just the initial value, so overwrite it
-          lastEmails = [];
-        }
-        return [...lastEmails, ...text.split(/\s+/g)];
-      });
-    });
-  }
-
-  function removeEmailByIndex(index) {
-    let newEmailsToAdd = [...emailsToAdd];
-    newEmailsToAdd.splice(index, 1);
-    setEmailsToAdd(newEmailsToAdd);
-  }
-
-  function removeResponseEmail(email) {
-    setEmailsToAdd(emailsToAdd.filter(item => item !== email));
-    setResponse(lastResponse => ({
-      ...lastResponse,
-      progress: lastResponse.progress.filter(email_obj => email_obj.email !== email)
-    }));
-  }
-
-  function handleAddStudentSubmit(e) {
-    e.preventDefault(); // prevent refresh on submit, which stops this request
-    setAddStage(CoordinatorAddStudentModal.STATES.LOADING);
-
-    let request_emails = emailsToAdd.map(email => {
-      let action = {};
-      if (responseActions.has(email)) {
-        action = responseActions.get(email);
-      }
-      return { email: email, ...action };
-    });
-
-    if (request_emails.length == 0) {
-      // no student emails to add, so just exit without reloading
-      closeModal();
-      return;
-    }
-
-    let request = { emails: request_emails, actions: {} };
-    if (responseActions.has("capacity")) {
-      request.actions["capacity"] = responseActions.get("capacity");
-    }
-
-    fetchWithMethod(`sections/${sectionId}/students/`, HTTP_METHODS.PUT, request).then(response => {
-      if (!response.ok) {
-        if (response.status === 500) {
-          // internal error
-          response.text().then(() => {
-            setResponse({
-              errors: { critical: `A internal error occurred. Please report this error to #tech-bugs immediately.` }
-            });
-            setAddStage(CoordinatorAddStudentModal.STATES.ERROR);
-          });
-        } else {
-          response.json().then(body => {
-            setResponse(body);
-            if (body.errors.critical) {
-              setAddStage(CoordinatorAddStudentModal.STATES.ERROR);
-            } else {
-              setAddStage(CoordinatorAddStudentModal.STATES.WARNING);
-            }
-          });
-        }
-        setValidationEnabled(false); // reset validation
-      } else {
-        // close modal and refresh the page
-        closeModal(true);
-      }
-    });
-  }
-
-  function updateResponseAction(e, email, field) {
-    let newAction = {};
-    if (responseActions.has(email)) {
-      newAction = responseActions.get(email);
-    }
-
-    if (e.target.checked) {
-      newAction[field] = e.target.value;
-    } else {
-      delete newAction[field];
-    }
-
-    let newResponseActions = new Map(responseActions);
-    newResponseActions.set(email, newAction);
-    setResponseActions(newResponseActions);
-  }
-
-  function updateGeneralResponseAction(e, field) {
-    let newResponseActions = new Map(responseActions);
-    if (e.target.checked) {
-      newResponseActions.set(field, e.target.value);
-    } else {
-      newResponseActions.delete(field);
-    }
-    setResponseActions(newResponseActions);
-  }
-
-  const initial_component = (
-    <React.Fragment>
-      <h4>Add new students</h4>
-      <div className="coordinator-email-content">
-        <div className="coordinator-email-input-list">
-          {emailsToAdd.map((email, index) => (
-            <div className="coordinator-email-input-item" key={index}>
-              <span className="inline-plus-sign ban-cancel" onClick={() => removeEmailByIndex(index)}>
-                ×
-              </span>
-              <input
-                className={"coordinator-email-input" + (validationEnabled ? "" : " lock-validation")}
-                type="email"
-                required
-                pattern=".+@berkeley.edu$"
-                title="Please enter a valid @berkeley.edu email address"
-                placeholder="New student's email"
-                list="user-emails-list"
-                value={email}
-                onChange={({ target: { value } }) => setEmail(index, value)}
-              />
-            </div>
-          ))}
-          <datalist id="user-emails-list">
-            {userEmails.map(email => (
-              <option key={email} value={email} />
-            ))}
-          </datalist>
-        </div>
-      </div>
-      <div className="coordinator-email-input-buttons">
-        <button className="coordinator-email-input-add" onClick={() => addNewEmail()}>
-          Add email
-        </button>
-        {
-          /* Firefox doesn't support clipboard reads from sites; readText would be undefined */
-          navigator.clipboard.readText && (
-            <button className="coordinator-email-input-add" onClick={() => addEmailsFromClipboard()}>
-              Add from clipboard
-            </button>
-          )
-        }
-        <button className="coordinator-email-input-submit" type="submit" onClick={() => setValidationEnabled(true)}>
-          Submit
-        </button>
-      </div>
-    </React.Fragment>
-  );
-
-  const loading_component = (
-    <React.Fragment>
-      <h4>Add new students</h4>
-      <LoadingSpinner />
-      <div></div> {/* empty element to align the content */}
-    </React.Fragment>
-  );
-
-  const ADD_STATUS = {
-    OK: "OK",
-    CONFLICT: "CONFLICT",
-    BANNED: "BANNED"
-  };
-
-  let ok_arr = [];
-  let conflict_arr = [];
-  let banned_arr = [];
-
-  if (response && response.progress) {
-    for (let email_obj of response.progress) {
-      switch (email_obj.status) {
-        case ADD_STATUS.OK:
-          ok_arr.push(email_obj);
-          break;
-        case ADD_STATUS.CONFLICT:
-          conflict_arr.push(email_obj);
-          break;
-        case ADD_STATUS.BANNED:
-          banned_arr.push(email_obj);
-          break;
-      }
-    }
-  }
-
-  const warning_component = (
-    <React.Fragment>
-      <h4>Add new students</h4>
-      <div className="coordinator-email-content">
-        <div className="coordinator-email-response-list">
-          {conflict_arr.length > 0 && (
-            <div className="coordinator-email-response-container">
-              <div className="coordinator-email-response-head">
-                <div className="coordinator-email-response-head-left coordinator-email-response-status-conflict">
-                  <ErrorCircle className="coordinator-email-response-status-conflict-icon" />
-                  Section conflict
-                </div>
-                <div className="coordinator-email-response-head-right">
-                  <div className="coordinator-email-reaponse-head-right-item">Drop?</div>
-                </div>
-              </div>
-              <div className="coordinator-email-response-item-container">
-                {conflict_arr.map(email_obj => {
-                  let conflictDetail = "";
-                  let drop_disabled = false;
-                  if (!email_obj.detail.section) {
-                    // look at reason
-                    if (!email_obj.detail.reason || email_obj.detail.reason === "other") {
-                      // unknown reason
-                      conflictDetail = "Unable to enroll user in section!";
-                    } else if (email_obj.detail.reason === "coordinator") {
-                      conflictDetail = "User is already a coordinator for the course!";
-                    } else if (email_obj.detail.reason === "mentor") {
-                      conflictDetail = "User is already a mentor for the course!";
-                    }
-                    drop_disabled = true;
-                  } else if (email_obj.detail.section.id == sectionId) {
-                    conflictDetail = "Already enrolled!";
-                    drop_disabled = true;
-                  } else {
-                    conflictDetail = (
-                      <React.Fragment>
-                        Conflict:{" "}
-                        <Link to={`/sections/${email_obj.detail.section.id}`} target="_blank" rel="noopener noreferrer">
-                          {email_obj.detail.section.mentor.name}
-                        </Link>
-                      </React.Fragment>
-                    );
-                  }
-                  return (
-                    <div key={email_obj.email} className="coordinator-email-response-item">
-                      <div className="coordinator-email-response-item-left">
-                        <div className="coordinator-email-response-item-left-email">
-                          <span
-                            className="inline-plus-sign ban-cancel"
-                            onClick={() => removeResponseEmail(email_obj.email)}
-                          >
-                            ×
-                          </span>
-                          {email_obj.email}
-                        </div>
-                        <div className="coordinator-email-response-item-left-detail">{conflictDetail}</div>
-                      </div>
-                      <div className="coordinator-email-response-item-right">
-                        <input
-                          type="checkbox"
-                          value="DROP"
-                          disabled={drop_disabled}
-                          onChange={e => updateResponseAction(e, email_obj.email, "conflict_action")}
-                        />
-                      </div>
-                    </div>
-                  );
-                })}
-              </div>
-            </div>
-          )}
-          {banned_arr.length > 0 && (
-            <div className="coordinator-email-response-container">
-              <div className="coordinator-email-response-head">
-                <div className="coordinator-email-response-head-left coordinator-email-response-status-banned">
-                  <ErrorCircle className="coordinator-email-response-status-conflict-icon" />
-                  Student banned
-                </div>
-                <div className="coordinator-email-response-head-right">
-                  <div className="coordinator-email-reaponse-head-right-item">
-                    Unban,
-                    <br /> Enroll
-                  </div>
-                  <div className="coordinator-email-reaponse-head-right-item">
-                    Unban,
-                    <br /> No Enroll
-                  </div>
-                </div>
-              </div>
-              <div className="coordinator-email-response-item-container">
-                {banned_arr.map(email_obj => (
-                  <div key={email_obj.email} className="coordinator-email-response-item">
-                    <div className="coordinator-email-response-item-left">
-                      <div className="coordinator-email-response-item-left-email">
-                        <span
-                          className="inline-plus-sign ban-cancel"
-                          onClick={() => removeResponseEmail(email_obj.email)}
-                        >
-                          ×
-                        </span>
-                        {email_obj.email}
-                      </div>
-                    </div>
-                    <div className="coordinator-email-response-item-right">
-                      <input
-                        type="radio"
-                        name={`unban-${email_obj.email}`}
-                        value="UNBAN_ENROLL"
-                        onChange={e => updateResponseAction(e, email_obj.email, "ban_action")}
-                      />
-                      <input
-                        type="radio"
-                        name={`unban-${email_obj.email}`}
-                        value="UNBAN_SKIP"
-                        onChange={e => updateResponseAction(e, email_obj.email, "ban_action")}
-                      />
-                    </div>
-                  </div>
-                ))}
-              </div>
-            </div>
-          )}
-          {ok_arr.length > 0 && (
-            <div className="coordinator-email-response-container">
-              <div className="coordinator-email-response-head">
-                <div className="coordinator-email-response-head-left coordinator-email-response-status-ok">
-                  <CheckCircle className="coordinator-email-response-status-ok-icon" />
-                  OK
-                </div>
-                <div className="coordinator-email-response-head-right"></div>
-              </div>
-              <div className="coordinator-email-response-item-container">
-                {ok_arr.map(email_obj => (
-                  <div key={email_obj.email} className="coordinator-email-response-item">
-                    <div className="coordinator-email-response-item-left">
-                      <span
-                        className="inline-plus-sign ban-cancel"
-                        onClick={() => removeResponseEmail(email_obj.email)}
-                      >
-                        ×
-                      </span>
-                      {email_obj.email}
-                    </div>
-                    <div className="coordinator-email-response-item-right"></div>
-                  </div>
-                ))}
-              </div>
-            </div>
-          )}
-          {response && response.errors && response.errors.capacity && (
-            <div className="coordinator-email-response-capacity-container">
-              <hr className="coordinator-email-response-hr" />
-              <div className="coordinator-email-response-email-status">
-                <div className="coordinator-email-response-capacity">
-                  <ErrorCircle className="coordinator-email-response-status-conflict-icon" />
-                  Section capacity exceeded!
-                </div>
-              </div>
-              <div className="coordinator-email-response-form">
-                <label>
-                  <input
-                    type="radio"
-                    name="capacity"
-                    value="EXPAND"
-                    onChange={e => updateGeneralResponseAction(e, "capacity")}
-                  />
-                  Enroll and expand section
-                </label>
-                <label>
-                  <input
-                    type="radio"
-                    name="capacity"
-                    value="SKIP"
-                    onChange={e => updateGeneralResponseAction(e, "capacity")}
-                  />
-                  Ignore (delete students manually)
-                </label>
-              </div>
-            </div>
-          )}
-        </div>
-      </div>
-      <div className="coordinator-email-input-buttons">
-        <button className="coordinator-email-input-submit" type="submit" onClick={() => setValidationEnabled(true)}>
-          Retry
-        </button>
-      </div>
-    </React.Fragment>
-  );
-
-  const error_component = (
-    <React.Fragment>
-      <h4 className="internal-error-title">Internal Error</h4>
-      <div className="internal-error-body">{response && response.errors && response.errors.critical}</div>
-      <div>Timestamp: {`${new Date()}`}</div>
-    </React.Fragment>
-  );
-
-  let curContents = "";
-  if (addStage === CoordinatorAddStudentModal.STATES.INITIAL) {
-    curContents = initial_component;
-  } else if (addStage === CoordinatorAddStudentModal.STATES.LOADING) {
-    curContents = loading_component;
-  } else if (addStage === CoordinatorAddStudentModal.STATES.ERROR) {
-    curContents = error_component;
-  } else if (addStage === CoordinatorAddStudentModal.STATES.WARNING) {
-    curContents = warning_component;
-  }
-
-  return (
-    <Modal className="coordinator-add-student-modal" closeModal={() => closeModal()}>
-      <form
-        id="coordinator-add-student-form"
-        className="coordinator-add-student-modal-contents"
-        onSubmit={handleAddStudentSubmit}
-      >
-        {curContents}
-      </form>
-    </Modal>
-  );
-}
-
-Object.defineProperty(CoordinatorAddStudentModal, "STATES", {
-  enumerable: false,
-  configurable: false,
-  writable: false,
-  value: Object.freeze({ INITIAL: "INITIAL", WARNING: "WARNING", ERROR: "ERROR", LOADING: "LOADING" })
-});
-
-CoordinatorAddStudentModal.propTypes = {
-  closeModal: PropTypes.func.isRequired,
-  userEmails: PropTypes.array.isRequired,
-  sectionId: PropTypes.number.isRequired
 };
 
 function StudentDropper({ id, name, reloadSection }) {
