+++ csm_web/frontend/src/images.d.ts
@@ -1,0 +1,6 @@
+/* Declare additional modules to allow imports of images */
+declare module "*.png";
+declare module "*.jpg";
+declare module "*.jpeg";
+declare module "*.svg";
+declare module "*.gif";
